+++ src/components/button-more.js
@@ -1,0 +1,3 @@
+export const createButtonMoreTemplate = () => (
+  `<button class="films-list__show-more">Show more</button>`
+);
